Global Refactoring (Delete everlasting comments, removing some cout) A lot of corrections about TFSM with guards Some testing with actual papers experimentation

+++ src/modules/model/algorithm/timedinputsequence.h
@@ -7,12 +7,12 @@ class TimedInputSequence : public Sequence
 {
 public:
     TimedInputSequence();
-    std::vector<std::pair<std::string, int> > content;
+    std::vector<std::pair<std::string, double> > content;
     std::string toString();
     int getSize();
-    std::pair<std::string, int> getElement(int index);
-    void addElement(std::pair<std::string, int> element);
-    void addElements(std::vector<std::pair<std::string, int> > elements);
+    std::pair<std::string, double> getElement(int index);
+    void addElement(std::pair<std::string, double> element);
+    void addElements(std::vector<std::pair<std::string, double> > elements);
 };
 
 #endif // TIMEDINPUTSEQUENCE_H
