+++ src/modules/model/machine/distinguishingautomaton_tfsm.h
@@ -12,7 +12,7 @@ public:
     void insertState(ProductState * state, std::string i, Guard g, ProductState * newState, bool isTimeout, int id);
     std::vector<executingPath> revealingPaths(Sequence * alpha);
     bool isPathDeterministic(const executingPath p);
-    void revealingPathsRecursive(ProductState * state, executingPath currentPath, std::vector<executingPath> & results, Sequence * alpha, int sequenceIndex, int timeBuffer, int timeLeftOver);
+    void revealingPathsRecursive(ProductState * state, executingPath currentPath, std::vector<executingPath> & results, Sequence * alpha, int sequenceIndex, double timeBuffer, double timeLeftOver);
     void print();
     Sequence * inputSequenceFromAcceptedLanguage(std::set<std::string> beginningStates, Sequence * prefix);
     void reachableStates(ProductState * state, executingPath currentPath, std::set<std::string> *results, Sequence * alpha, int sequenceIndex, int timeBuffer, int timeLeftOver);
