+++ src/modules/model/structs.h
@@ -6,7 +6,7 @@
 #include <string>
 #include "machine/guard.h"
 
-typedef std::pair<std::string, int> ts;
+typedef std::pair<std::string, double> ts;
 typedef std::pair<std::string, Guard> its;
 //typedef std::vector<ts> sequence;
 typedef std::vector<int> executingPath;
